refactor CompactionTaskComparator
